Fix pseudo markup storage cleanup

+++ src/mlsystem2/training_ui_api/_models.py
@@ -6,7 +6,7 @@ import uuid
 from datetime import datetime
 from typing import Any
 
-from sqlalchemy import Boolean, DateTime, ForeignKey, Index, Integer, String, Text, UniqueConstraint, Uuid, func
+from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Index, Integer, String, Text, UniqueConstraint, Uuid, func
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy.types import JSON
@@ -66,7 +66,7 @@ class StoredFileRow(Base):
     original_name: Mapped[str] = mapped_column(String(512))
     content_type: Mapped[str | None] = mapped_column(String(160), nullable=True)
     path: Mapped[str] = mapped_column(Text)
-    size_bytes: Mapped[int] = mapped_column(Integer)
+    size_bytes: Mapped[int] = mapped_column(BigInteger)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
